+++ cfdm/cfdmimplementation.py
@@ -424,7 +424,7 @@ class CFDMImplementation(Implementation):
     def get_cell_connectivities(self, parent):
         """Return the cell connectivities from a parent.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters:
 
@@ -1034,9 +1034,9 @@ class CFDMImplementation(Implementation):
         return field.domain_axes(todict=True)[axis].get_size()
 
     def get_domain_topologies(self, parent):
-        """Return the domain topologies from  a parent.
+        """Return the domain topologies from a parent.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters:
 
@@ -2029,9 +2029,9 @@ class CFDMImplementation(Implementation):
         :Parameters:
 
             kwargs: optional
-                Parameters with which to intialising the object.
+                Parameters with which to initialise the object.
 
-                .. versionadded:: (cfdm) NEXTVERSION
+                .. versionadded:: (cfdm) 1.13.1.0
 
         :Returns:
 
@@ -2627,7 +2627,7 @@ class CFDMImplementation(Implementation):
             kwargs: optional
                 Initialisation parameters to pass to the new instance.
 
-                .. versionadded:: (cfdm) NEXTVERSION
+                .. versionadded:: (cfdm) 1.13.1.0
 
         :Returns:
 
@@ -2640,7 +2640,7 @@ class CFDMImplementation(Implementation):
     def initialise_PyfiveArray(self, **kwargs):
         """Return a `PyfiveArray` instance.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters:
 
@@ -3661,7 +3661,7 @@ class CFDMImplementation(Implementation):
             if data is not None:
                 filenames += tuple(data._original_filenames())
 
-        parent._original_filenames(define=set(filenames))
+        parent._original_filenames(define=filenames)
 
     def set_parameter(self, parent, parameter, value, copy=True):
         """Set a parameter on a component.