+++ cfdm/data/netcdf4array.py
@@ -33,7 +33,7 @@ class NetCDF4Array(IndexMixin, abstract.FileArray):
         from the netCDF variable *var* and stored for fast future
         access.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters: